mr-owlf-dss/requirements.txt
cassandra-driver==3.19.0
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
six==1.12.0
